fix: add ignore to archived and forked repositories on search (#56)
Many orgs have multiple repos that are forks and archived. adding
`isArchived: false` and `isFork: false` to the `repositories` query
reduces the overhead of getting metrics on non-maintained repositories
and increases performance by reducing the amount of subsequent calls and
iterations. This may be feature flagged via config in the future.