feat: add ability to perform a search query #68 (#70)
Adds the ability to use a search query to get the repositories needed for metrics instead of relying on the org as a whole all the time. For examples, see the config example in the readme. Additionally this cleans up some code, updates the OTEL library and builder versions to `v0.83.0` and correspondingly updates the go version to `1.20`.