feat: move to singular search query for repos and fix branch age queries
Remove the separate instance of queries based on organization 
or user in favor of the repo search query. This removes a significant 
amount of generated code and reduces logical overhead. Additionally, 
the branch information query has been refactored into the new
query format for better readability and performance. A bug was fixed 
with those queries not getting true information and an index out of range bug was also fixed.