feat: added branchCount metric
This commit adds support for emitting the amount of branches in a given repo. There is a caveat that it only supports up to 100,000 branches so far. The official GitLab group on GitLab has a repo with well over 100,000 branches we can use as a test case but I had a hard time finding others with this many. Either way, we need to make sure we handle this case.