feat: add gitlab branches count metric (#97)
* refactor: updated branchNames max to 100,000
As discussed in a Huddle with @adrielp, this commit updates the limit of branchNames returned from the GitLab GraphQL API to 100,000 since it doesn't seem to have a hard limit and 100,000 covers the _vast_ majority of cases without hitting performance too hard. Additional work will be done to handle this more appropriately in the future. 

* refactor: added pagination info for projects graphql query

* refactor: added support to get all projects

* docs: updated README files with gitlab info

* feat: add go routines to the branch count retrieval to improve performance

* ci: added markdownlint pre-commit

---------

Signed-off-by: Devin W. Leaman <git@4lch4.email>
Signed-off-by: Adriel Perkins <adrielp@liatrio.com>
Co-authored-by: Joe Wood <joew@liatrio.com>
Co-authored-by: Adriel Perkins <adrielp@liatrio.com>