feat(githubscraper): add rate limiter
This commit adds a rate limiter to the GitHub scraper. The rate limiter
is implemented as a Strategy Pattern. So other gitproviders can
handle their own rate limiting logic.

This commit also adds a client wrapper that implements the Doer
interface so we can hook into request/response headers and ultimately
the rate limiting logic.