feat: upgrade receiver versions and update renovate config to file matcher (#262)
Updates the `renovate.json` to use file match managers. Removed the automerge which will come back in a subsequent pull request after verification of fixes to Renovate. 

Updates the dependencies on the components to bring them up-to-date with the latest from the OTEL and other projects. 

After updating, adjusts the gitlab scraper to leverage `gitlab.Ptr` in place of its deprecated alternatives.